KOGITO-3545: dmn-tracing examples are now completed and aligned
