+++ dmn-tracing-quarkus/src/test/java/org/kie/dmn/kogito/quarkus/tracing/NativeLoanEligibilityIT.java
@@ -15,14 +15,13 @@
  */
 package org.kie.dmn.kogito.quarkus.tracing;
 
-import org.kie.kogito.testcontainers.quarkus.KafkaQuarkusTestResource;
-
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.NativeImageTest;
+import org.kie.kogito.testcontainers.quarkus.KafkaQuarkusTestResource;
 
 @NativeImageTest
 @QuarkusTestResource(KafkaQuarkusTestResource.class)
-public class NativeLoanEligibilityIT extends LoanEligibilityTest {
+public class NativeLoanEligibilityIT extends LoanEligibilityIT {
 
     // Execute the same tests but in native mode.
 }
