+++ dmn-tracing-springboot/operator/dmn-tracing-springboot.yaml
@@ -1,0 +1,23 @@
+apiVersion: app.kiegroup.org/v1alpha1
+kind: KogitoBuild
+metadata:
+  name: dmn-springboot-example
+spec:
+  type: RemoteSource
+  runtime: springboot
+  #env:
+  # env can be used to set variables during build
+  #- name: MY_CUSTOM_ENV
+  #  value: "my value"
+  gitSource:
+    contextDir: dmn-springboot-example
+    uri: 'https://github.com/kiegroup/kogito-examples'
+  # set your maven nexus repository to speed up the build time
+  #mavenMirrorURL: 
+---
+apiVersion: app.kiegroup.org/v1alpha1
+kind: KogitoRuntime
+metadata:
+  name: dmn-springboot-example
+spec:
+  runtime: springboot
