+++ dmn-tracing-springboot/src/main/resources/application.properties
@@ -1,3 +1,3 @@
 server.address=0.0.0.0
 resteasy.jaxrs.scan-packages=org.kie.kogito.**,http*
-kafka.bootstrap.address=localhost:9092
+kogito.addon.tracing.decision.kafka.bootstrapAddress=localhost:9092
