--- dmn-tracing-springboot/src/test/java/org/kie/dmn/kogito/springboot/tracing/LoanEligibilityTest.java
@@ -1,76 +1,0 @@
-/**
- *  Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-package org.kie.dmn.kogito.springboot.tracing;
-
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.KafkaContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.Matchers.is;
-
-@Testcontainers
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, classes = KogitoSpringbootApplication.class)
-public class LoanEligibilityTest {
-
-    @Container
-    public static KafkaContainer kafkaContainer = new KafkaContainer();
-
-    @DynamicPropertySource
-    static void kafkaProperties(DynamicPropertyRegistry registry) {
-        registry.add("kafka.bootstrap.address", kafkaContainer::getBootstrapServers);
-    }
-
-    @LocalServerPort
-    private int port;
-
-    @BeforeEach
-    public void setUp() {
-        RestAssured.port = port;
-    }
-
-    @Test
-    public void testEvaluateLoanEligibility() {
-        given()
-                .body("{" +
-                      "    \"Client\": {" +
-                      "        \"age\": 43," +
-                      "        \"salary\": 1950," +
-                      "        \"existing payments\": 100" +
-                      "    }," +
-                      "    \"Loan\": {" +
-                      "        \"duration\": 15," +
-                      "        \"installment\": 180" +
-                      "    }," +
-                      "    \"SupremeDirector\" : \"Yes\"," +
-                      "    \"Bribe\": 1000" +
-                      "}")
-                .contentType(ContentType.JSON)
-            .when()
-                .post("/LoanEligibility")
-            .then()
-                .statusCode(200)
-                .body("'Decide'", is(true));
-    }
-}
